# Optimize ColumnValueSegmentPruner by caching value hashes