@gortiz 2022-05-25
The same content as before, but all together, sorted by number of segments:

```
Benchmark  (hash/iden)  (_numSegments)    Score    Error  Units
V1               iden         10          0.972  ± 0.003  us/op
V0               iden         10          1.426  ± 0.003  us/op
V2          hash              10          1.325  ± 0.029  us/op
V2               iden         10          1.110  ± 0.001  us/op
V3          hash              10          0.971  ± 0.003  us/op
V3               iden         10          0.556  ± 0.001  us/op

V0               iden        100         14.719  ± 1.342  us/op
V1               iden        100         10.149  ± 0.046  us/op
V2          hash             100         12.851  ± 0.171  us/op
V2               iden        100         10.659  ± 0.139  us/op
V3          hash             100          4.843  ± 0.045  us/op
V3               iden        100          4.964  ± 0.268  us/op

V0               iden       1000        174.879  ± 5.115  us/op
V1               iden       1000        135.063  ±14.670  us/op
V2          hash            1000        138.314  ± 2.238  us/op
V2               iden       1000        120.608  ± 0.159  us/op
V3          hash            1000         72.658  ± 0.735  us/op
V3               iden       1000         58.304  ± 0.382  us/op
```